p2p-fetch API revamp to prevent creating multiple ServiceWorkers
Previously the parameters were passed via URL and any change in those
resulted in *another* Service Worker being created (instead of reusing
the existing one).
With this new API a single ServiceWorker is created and it's
configuration is passed in via postMessage. This offers the ability to
update the configuration in real time!